Added persistent/onplayerload option.

+++ client/client.lua
@@ -1,4 +1,20 @@
 local entity_ids = {}
+if Config.Persistent then
+    if Config.Framework == 'qb' then
+        QBCore = exports['qb-core']:GetCoreObject()
+        AddEventHandler('QBCore:Client:OnPlayerLoaded', function()
+            Wait(1500)
+            SpawnProps()
+        end)
+    elseif Config.Framework == 'esx' then
+        local ESX = exports["es_extended"]:getSharedObject()
+        RegisterNetEvent('esx:playerLoaded')
+        AddEventHandler('esx:playerLoaded', function()
+            Wait(1500)
+            SpawnProps()
+        end)
+    end
+end
 
 function SpawnProps()
     for k, v in pairs(Config.PropLocations) do
@@ -6,9 +22,9 @@ function SpawnProps()
         while not HasModelLoaded(v.model) do
             Wait(0)
         end
-        local prop = CreateObject(v.model, v.coords.x, v.coords.y, v.coords.z - 1, true, true, false)
+        local prop = CreateObject(v.model, v.coords.x, v.coords.y, v.coords.z - 1, false, true, false)
         entity_ids[#entity_ids + 1] = prop
-        SetEntityHeading(prop, v.coords.w -180)
+        SetEntityHeading(prop, v.coords.w - 180)
         FreezeEntityPosition(prop, true)
         print("^3Prop Spawned:" .. v.coords)
     end
@@ -22,13 +38,14 @@ function DeleteProps()
     end
 end
 
-RegisterCommand(Config.DeleteCommand, function()
-    DeleteProps()
-end)
-
-RegisterCommand(Config.SpawnCommand, function()
-    SpawnProps()
-end)
+if Config.debug then
+    RegisterCommand(Config.DeleteCommand, function()
+        DeleteProps()
+    end)
+    RegisterCommand(Config.SpawnCommand, function()
+        SpawnProps()
+    end)
+end
 
 AddEventHandler('onResourceStop', function(resourceName)
     if (GetCurrentResourceName() ~= resourceName) then return end
